kafka: refactor the kafka sink further (#772)
* enable the kafka sink metrics

* tiny adjust

* pass ctx directly

* set metrics collector to the kafka sink

* move all to kafka sink factory

* refactor the kafka sink

* fix format

* some adjust

* set the error group

* add ctx

* tiny refactor the kafka sink

* refactor the kafka factory interface

* do not return error when close resource.

* rename producer

* remove unused field

* adjust kafka sink run

* adjust the format

* fix the make check

+++ pkg/sink/codec/avro/schema_manager.go
@@ -24,7 +24,7 @@ type SchemaManager interface {
 	Register(ctx context.Context, schemaName string, schemaDefinition string) (schemaID, error)
 	Lookup(ctx context.Context, schemaName string, schemaID schemaID) (*goavro.Codec, error)
 	GetCachedOrRegister(ctx context.Context, topicName string,
-		tableVersion uint64, schemaGen SchemaGenerator) (*goavro.Codec, []byte, error)
+		tableVersion uint16, schemaGen SchemaGenerator) (*goavro.Codec, []byte, error)
 	RegistryType() string
 	ClearRegistry(ctx context.Context, schemaName string) error
 }
